server/requirements.txt
fastapi
uvicorn
sqlmodel
requests
pydantic-settings
pydantic
pydantic_core
python-dotenv
pymysql
cryptography
passlib[bcrypt]
bcrypt==3.2.2
python-jose[cryptography]
google-auth-oauthlib
google-api-python-client
mysql-connector-python
alembic